+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed:
+- Health chart legend on select
+
+## [0.8.151] - 2022-09-22
 ### Changed:
 - Style: Inconsolata as monospace font
 - New color theme in journal and tasks list